Smarter way to compute the forces
Compute the field at the end of the time-step. This will get used at the next time-step.
The change from the previous version is that sometimes I was computing the same force again because of
is_first_step flag. So technically no difference to the results; just avoiding one force calculation call.